pyproject.toml
[tool.poetry]
name = "python-fastapi-cookbook"
version = "0.1.0"
description = ""
authors = ["Benny Cheng <benny_cheng@leadtek.com.tw>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = { extras = ["all"], version = "^0.115.8" }
uvicorn = "^0.34.0"
sqlalchemy = "^2.0"
pydantic = { extras = ["email"], version = "^2.10.6" }
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
loguru = "^0.7.3"
update = "^0.0.1"
httpx = "^0.28.1"
pyotp = "^2.9.0"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.1.2"
isort = "^6.0.0"
mypy = "^1.15.0"
pre-commit = "^4.1.0"
pytest = "^8.3.4"
commitizen = "^4.3.0"

[tool.black]
line-length = 100
skip-string-normalization = false
target-version = ["py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
src_paths = ["src"]

[tool.mypy]
python_version = "3.11"
# strict = true
# disallow_untyped_defs = true
# disallow_untyped_calls = true
# disallow_any_generics = true
ignore_missing_imports = true
# warn_return_any = true
# warn_unused_configs = true

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver"
version = "0.0.1"
update_changelog_on_bump = true
major_version_zero = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
